refactor: Move methods from `others` package to their respective packages
